AV-52543 AV-55613 AV-55707 AV-55706 AV-55704 Refactored code and removed abstract class. ApiSession class can handle SAML authentications with support of IDP classes.
