+++ python/avi/sdk/saml_avi_api.py
@@ -15,103 +15,6 @@ import logging
 logger = logging.getLogger(__name__)
 
 
-def get_idp_class(idp):
-    """
-    Returns specific IDP class as
-    per the idp input.
-    :param idp: IDP type such as okta, onelogin, etc
-    :param args:
-    :param kwargs:
-    :return: idp class
-    """
-    if not idp:
-        logger.error("Please provide IDP name.")
-        raise StandardError("Please provide IDP name")
-    if str(idp).lower() == "okta":
-        logger.info("Using OktaSAMLApiSession to create session")
-        return OktaSAMLApiSession
-    if str(idp).lower() == "onelogin":
-        logger.info("Using OneloginSAMLApiSession to create session")
-        return OneloginSAMLApiSession
-
-
-class SAMLApiSession(object):
-    """
-    This creates corresponding IDP class as per the idp input
-    and implement SAML authentication and create controller session.
-    """
-    # At runtime creates a specific IDP class object
-    # before initialization
-    def __new__(cls, *args, **kwargs):
-        idp = kwargs.get("idp", None)
-        idp_class = get_idp_class(idp)
-        return idp_class(*args, **kwargs)
-
-    @staticmethod
-    def get_session(
-            controller_ip=None, username=None, password=None, token=None,
-            tenant=None, tenant_uuid=None, verify=False, port=None,
-            timeout=60, api_version=None, retry_conxn_errors=True,
-            data_log=False, avi_credentials=None, session_id=None,
-            csrftoken=None, lazy_authentication=False,
-            max_api_retries=None, idp_cookies=None, idp=None):
-        """
-        returns the session object for same user and tenant
-        calls init if session dose not exist and adds it to session cache
-        :param controller_ip: Controller Ip
-        :param username: IDP username
-        :param password: IDP password
-        :param verify: Verify SSL boolean flag
-        :param tenant: Name of the tenant on Avi Controller
-        :param tenant_uuid: Don't specify tenant when using tenant_id
-        :param port: Rest-API may use a different port other than 443
-        :param timeout: timeout for API calls; Default value is 60 seconds
-        :param retry_conxn_errors: retry on connection errors
-        :param api_version: Controller API version
-        :param idp_cookies: Identity Provider cookies.
-        :param idp: IDP name such as  onelogin, okta, etc
-        """
-        if not avi_credentials:
-            tenant = tenant if tenant else "admin"
-            avi_credentials = AviCredentials(controller=controller_ip,
-                                             username=username,
-                                             password=password,
-                                             token=token,
-                                             api_version=api_version,
-                                             tenant=tenant,
-                                             tenant_uuid=tenant_uuid,
-                                             port=port,
-                                             timeout=timeout,
-                                             session_id=session_id,
-                                             csrftoken=csrftoken)
-        k_port = avi_credentials.port if avi_credentials.port else 443
-        if avi_credentials.controller.startswith('http'):
-            k_port = 80 if not avi_credentials.port else k_port
-        key = '%s:%s:%s' % (avi_credentials.controller,
-                            avi_credentials.username, k_port)
-        cached_session = sessionDict.get(key)
-        if cached_session:
-            user_session = cached_session['api']
-            if not (user_session.avi_credentials.csrftoken or
-                    lazy_authentication):
-                user_session.authenticate_session()
-        else:
-            idp_class = get_idp_class(idp)
-            user_session = idp_class(
-                controller=controller_ip, username=username,
-                password=password, token=token,
-                tenant=tenant, tenant_uuid=tenant_uuid,
-                verify=verify, port=port, timeout=timeout,
-                retry_conxn_errors=retry_conxn_errors,
-                api_version=api_version, data_log=data_log,
-                avi_credentials=avi_credentials,
-                lazy_authentication=lazy_authentication,
-                max_api_retries=max_api_retries,
-                idp_cookies=idp_cookies, idp=idp,
-            )
-        return user_session
-
-
 class OneloginSAMLApiSession(ApiSession):
     """"
         Extends the ApiSession class to override authentication
